+++ src/features/offer-sorting-select/index.ts
@@ -1,0 +1,7 @@
+import { OfferSortingSelect } from './ui/offer-sorting-select';
+import sortingSliceReducer from './model/offer-sorting-select-slice';
+import { offerSortTypeToComparerMap } from './lib/sorting-functions';
+import { useOfferSorting } from './lib/use-offer-sorting';
+import { OfferSortType } from './model/types';
+
+export { OfferSortingSelect , sortingSliceReducer, offerSortTypeToComparerMap, useOfferSorting, OfferSortType };
