+++ src/features/offer-sorting-select/lib/use-offer-sorting/index.ts
@@ -1,0 +1,3 @@
+import { useOfferSorting } from './use-offer-sorting';
+
+export { useOfferSorting };
