+++ src/features/offer-sorting-select/ui/offer-sorting-select/index.ts
@@ -1,0 +1,3 @@
+import { OfferSortingSelect } from './offer-sorting-select';
+
+export { OfferSortingSelect };
