--- src/features/price-sorting-select/index.ts
@@ -1,7 +1,0 @@
-import { PriceSortingSelect } from './ui/price-sorting-select';
-import sortingSliceReducer from './model/price-sorting-select-slice';
-import { offerSortTypeToComparerMap } from './lib/sorting-functions';
-import { usePriceSorting } from './lib/use-price-sorting';
-import { OfferSortType } from './model/types';
-
-export { PriceSortingSelect, sortingSliceReducer, offerSortTypeToComparerMap, usePriceSorting, OfferSortType };
