--- src/features/price-sorting-select/lib/use-price-sorting/index.ts
@@ -1,3 +1,0 @@
-import { usePriceSorting } from './use-price-sorting';
-
-export { usePriceSorting };
