--- src/features/price-sorting-select/ui/price-sorting-select/index.ts
@@ -1,3 +1,0 @@
-import { PriceSortingSelect } from './price-sorting-select';
-
-export { PriceSortingSelect };
